requirements.txt
flask
flask_cors
mysql-connector-python
jsonify
